Golden Eagle Flight Plan.
This project was developed during the Advanced Web development course
work. This project was developed using Java Spring and Hibernate with
PostgreSQL . In a most simple form, this is a checklist that helps
guiding the students through their college career. It looks like a table
in which columns are called as Runways, Rows are Stages and the contents
in between are checkpoints with checkboxes. All the three can be
added,updated and deleted upon that they are re-ordered by dragging them
and the checked checklists should remain the same till it is been
unchecked all these should happen without reloading when the functions
are made, this was implemented using Ajax. There are three types of
users students, advisor and dean.